+++ packages/tsurlfilter/test/rules/declarative-converter/filter.test.ts
@@ -4,8 +4,18 @@ import { Filter } from '../../../src/rules/declarative-converter';
 describe('Filter', () => {
     // NOTE: Testing filter SHOULD contain some of convertible while
     // preprocessing rules.
-    // eslint-disable-next-line max-len
-    const rawContent = '||example.com^$document\r\n||example.net^\r\n@@||example.io^\r\n||googletagmanager.com/gtm.js$script,xmlhttprequest,redirect=googletagmanager-gtm,domain=einthusan.ca|einthusan.tv|einthusan.com\r\n||googletagmanager.com/gtm.js$script,redirect=googletagmanager-gtm,domain=lastampa.it\r\nsamnytt.se#@#div[class=""], a, .sticky > div[style="display:grid"], .post-content > div.mx-auto:has-text(/annons/i)\r\nwolt.com##button:has(> div > div > div > span:has-text(Sponsored))\r\n4wank.com#?#.video-holder > center > :-abp-contains(/^Advertisement$/)';
+    const rawContent = [
+        '||example.com^$document',
+        '||example.net^',
+        '@@||example.io^',
+        // eslint-disable-next-line max-len
+        '||googletagmanager.com/gtm.js$script,xmlhttprequest,redirect=googletagmanager-gtm,domain=einthusan.ca|einthusan.tv|einthusan.com',
+        '||googletagmanager.com/gtm.js$script,redirect=googletagmanager-gtm,domain=lastampa.it',
+        // eslint-disable-next-line max-len
+        'samnytt.se#@#div[class=""], a, .sticky > div[style="display:grid"], .post-content > div.mx-auto:has-text(/annons/i)',
+        'wolt.com##button:has(> div > div > div > span:has-text(Sponsored))',
+        '4wank.com#?#.video-holder > center > :-abp-contains(/^Advertisement$/)',
+    ].join('\r\n');
 
     it('loads content from string source provider', async () => {
         const filter = new Filter(
@@ -36,5 +46,87 @@ describe('Filter', () => {
         const preprocessedContent = preprocessedFilter.rawFilterList.split('\r\n');
 
         expect(rules).toStrictEqual(preprocessedContent);
+    });
+
+    it('unloads content correctly', async () => {
+        const filter = new Filter(
+            1,
+            { getContent: async () => FilterListPreprocessor.preprocess(rawContent) },
+            true,
+        );
+
+        // Load content
+        await filter.getContent();
+
+        expect(Object.getOwnPropertyDescriptor(filter, 'content')?.value).not.toBeNull();
+
+        // Unload content
+        filter.unloadContent();
+        expect(Object.getOwnPropertyDescriptor(filter, 'content')?.value).toBeNull();
+    });
+
+    it('does not return stale content after unload', async () => {
+        const filter = new Filter(
+            1,
+            { getContent: async () => FilterListPreprocessor.preprocess(rawContent) },
+            true,
+        );
+
+        // Load content
+        await filter.getContent();
+
+        expect(Object.getOwnPropertyDescriptor(filter, 'content')?.value).not.toBeNull();
+
+        // Unload content
+        filter.unloadContent();
+        expect(Object.getOwnPropertyDescriptor(filter, 'content')?.value).toBeNull();
+
+        // Reload content after unloading
+        const newContent = await filter.getContent();
+        expect(FilterListPreprocessor.getOriginalFilterListText(newContent)).toStrictEqual(rawContent);
+    });
+
+    it('waits for content to load before unloading', async () => {
+        let resolveFetch: (content: any) => void;
+
+        const fetchPromise = new Promise((resolve) => {
+            resolveFetch = resolve;
+        });
+
+        const filter = new Filter(
+            1,
+            { getContent: async () => fetchPromise as any },
+            true,
+        );
+
+        // Start loading content
+        const contentPromise = filter.getContent();
+
+        // Call unloadContent while content is still loading
+        filter.unloadContent();
+
+        // Resolve content fetch
+        resolveFetch!(FilterListPreprocessor.preprocess(rawContent));
+
+        // Ensure content is still correctly unloaded after the fetch completes
+        await contentPromise;
+
+        expect(Object.getOwnPropertyDescriptor(filter, 'content')?.value).toBeNull();
+    });
+
+    it('prevents memory leaks by resetting contentLoadingPromise', async () => {
+        const filter = new Filter(
+            1,
+            { getContent: async () => FilterListPreprocessor.preprocess(rawContent) },
+            true,
+        );
+
+        await filter.getContent();
+
+        expect(Object.getOwnPropertyDescriptor(filter, 'contentLoadingPromise')?.value).toBeNull();
+
+        filter.unloadContent();
+
+        expect(Object.getOwnPropertyDescriptor(filter, 'contentLoadingPromise')?.value).toBeNull();
     });
 });